feat(Env): feat make sure env name unique
Closes #69

+++ src/app/setting/env-setting/env-setting.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import {GableBackendService} from '../../core/services/gable-backend.service';
 import {NzContextMenuService, NzDropdownMenuComponent} from 'ng-zorro-antd/dropdown';
 import {NzModalRef} from 'ng-zorro-antd/modal';
+import {TranslateService} from "@ngx-translate/core";
+import {NzMessageService} from "ng-zorro-antd/message";
 
 @Component({
   selector: 'app-env-setting',
@@ -22,6 +24,7 @@ export class EnvSettingComponent implements OnInit {
   isGettingData = false;
   constructor(private nzContextMenuService: NzContextMenuService,
               private gableBackendService: GableBackendService,
+              private messageService: NzMessageService,
               private modal: NzModalRef) { }
 
   ngOnInit(): void {
@@ -37,7 +40,8 @@ export class EnvSettingComponent implements OnInit {
 {
   "replace": {},
   "add": {},
-  "remove": {}
+  "remove": {},
+  "removeByIndex": {}
 }
       `
     };
@@ -66,16 +70,34 @@ export class EnvSettingComponent implements OnInit {
 
   saveConfig() {
     if (this.selectedId.startsWith('addNew_')) {
-      this.gableBackendService.addEnv(this.configStr, this.name, this.type).subscribe(res => {
-        localStorage.setItem('env', JSON.stringify(res.data));
-        this.gableBackendService.setEnv(res.data);
-        this.modal.destroy({});
+      let have = false;
+      this.envs.forEach(value => {
+        if (value.name === this.name) {
+          have = true;
+        }
       });
-    }else {
+      if (have) {
+        this.messageService.error('Env Name: ' + this.name + ' Have Exist', {nzDuration: 3500});
+        return;
+      }
+      this.gableBackendService.addEnv(this.configStr, this.name, this.type).subscribe(res => {
+        if (res.result) {
+          localStorage.setItem('env', JSON.stringify(res.data));
+          this.gableBackendService.setEnv(res.data);
+          this.modal.destroy({});
+        } else {
+          this.messageService.error('Save Env Error ' + res.message, {nzDuration: 3500});
+        }
+      });
+    } else {
       this.gableBackendService.updateEnv(this.configStr, this.name, this.selectedId).subscribe(res => {
-        localStorage.setItem('env', JSON.stringify(res.data));
-        this.gableBackendService.setEnv(res.data);
-        this.modal.destroy({});
+        if (res.result) {
+          localStorage.setItem('env', JSON.stringify(res.data));
+          this.gableBackendService.setEnv(res.data);
+          this.modal.destroy({});
+        } else {
+          this.messageService.error('Save Env Error ' + res.message, {nzDuration: 3500});
+        }
       });
     }
   }
